Updated scripts

+++ providers/protonvpn/vpn.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/env bash
 
+err() { echo -e "${color:+\e[31m}[!] $@\e[0m"; }
+
+errx() { echo -e "${color:+\e[31m}[!] ${@:2}\e[0m"; exit $1; }
+
 default_gateway() {
-    local ret
-    [[ -z $conf ]] || ret="$(cat $conf | jq -cMrS ".$vpn.gateway")"
-    [[ $ret != "null" ]] || unset ret
+    local ret="$(json_get gateway)"
     echo "${ret:-100}" # US-FREE-01
 }
 
@@ -12,29 +14,44 @@ get_gateway() {
     local index="0"
 
     case "$gateway_selection" in
-        "random") let "index = $RANDOM % ${#gateways[@]}" ;;
-        *) let "index = $(default_gateway) - 1" ;;
+        "random") index="$(( $RANDOM % ${#gateways[@]} ))" ;;
+        *) index="$(( $(default_gateway) - 1 ))" ;;
     esac
 
     echo "${gateways[$index]}"
+}
+
+good() { echo -e "${color:+\e[32m}[+] $@\e[0m"; }
+
+info() { echo -e "${color:+\e[37m}[*] $@\e[0m"; }
+
+json_get() {
+    [[ -z $conf ]] || jq -cMrS ".$vpn.$@" $conf | sed -r "s/^null$//g"
 }
 
 list_gateways() {
     find . -iname "*.ovpn" | sed -r "s#./|\.ovpn##g" | sort
 }
 
+long_opt() {
+    local arg shift="0"
+    case "$1" in
+        "--"*"="*) arg="${1#*=}"; [[ -n $arg ]] || usage 1 ;;
+        *) shift="1"; shift; [[ $# -gt 0 ]] || usage 1; arg="$1" ;;
+    esac
+    echo "$arg"
+    return $shift
+}
+
 setup_creds() {
     rm -f creds.txt
 
-    local password
-    local username
+    local password username
 
     if [[ -n $conf ]]; then
-        local creds="$(cat $conf | jq -cMrS ".$vpn.credsfile")"
-        local cfile="$(
-            cat $conf | jq -cMrS ".$vpn.encrypted_credsfile"
-        )"
-        if [[ $cfile != "null" ]]; then
+        local creds="$(json_get credsfile)"
+        local cfile="$(json_get encrypted_credsfile)"
+        if [[ -n $cfile ]]; then
             if [[ -f $confdir/$cfile ]]; then
                 gpg -dq $confdir/$cfile >creds.txt
                 chmod 400 creds.txt
@@ -44,9 +61,9 @@ setup_creds() {
                 chmod 400 creds.txt
                 return
             else
-                echo "$cfile does not exist"
+                warn "$cfile does not exist"
             fi
-        elif [[ $creds != "null" ]]; then
+        elif [[ -n $creds ]]; then
             if [[ -f $confdir/$creds ]]; then
                 cp -f $confdir/$creds creds.txt
                 chmod 400 creds.txt
@@ -56,48 +73,36 @@ setup_creds() {
                 chmod 400 creds.txt
                 return
             else
-                echo "$creds does not exist"
+                warn "$creds does not exist"
             fi
         else
-            password="$(cat $conf | jq -cMrS ".$vpn.password")"
-            local pfile="$(
-                cat $conf | jq -cMrS ".$vpn.encrypted_password"
-            )"
-            if [[ $pfile != "null" ]]; then
+            password="$(json_get password)"
+            local pfile="$(json_get encrypted_password)"
+            if [[ -n $pfile ]]; then
                 if [[ -f $confdir/$pfile ]]; then
                     password="$(gpg -dq $confdir/$pfile)"
                 elif [[ -f $pfile ]]; then
                     password="$(gpg -dq $pfile)"
                 else
-                    echo "$pfile does not exist"
+                    warn "$pfile does not exist"
                 fi
-            elif [[ $password == "null" ]]; then
-                unset password
             fi
-            username="$(cat $conf | jq -cMrS ".$vpn.username")"
-            local ufile="$(
-                cat $conf | jq -cMrS ".$vpn.encrypted_username"
-            )"
-            if [[ $ufile != "null" ]]; then
+            username="$(json_get username)"
+            local ufile="$(json_get encrypted_username)"
+            if [[ -n $ufile ]]; then
                 if [[ -f $confdir/$ufile ]]; then
                     username="$(gpg -dq $confdir/$ufile)"
                 elif [[ -f $ufile ]]; then
                     username="$(gpg -dq $ufile)"
                 else
-                    echo "$ufile does not exist"
+                    warn "$ufile does not exist"
                 fi
-            elif [[ $username == "null" ]]; then
-                unset username
             fi
         fi
     fi
 
-    if [[ -z $username ]]; then
-        read -p "Enter username: " username
-    fi
-    if [[ -z $password ]]; then
-        read -p "Enter password: " -s password
-    fi
+    [[ -n $username ]] || read -p "Enter username: " username
+    [[ -n $password ]] || read -p "Enter password: " -s password
 
     echo "$username" >creds.txt
     echo "$password" >>creds.txt
@@ -106,27 +111,24 @@ setup_creds() {
 
 start_vpn() {
     local gateway="$(get_gateway)"
-    echo "Using gateway: $gateway"
+    info "Using gateway: $gateway"
     setup_creds
     sudo openvpn $gateway.ovpn
     [[ $? -eq 0 ]] || stop_vpn
 }
 
 stop_vpn() {
-    echo -n "Killing process and cleaning up..."
-    if [[ -n $(pgrep openvpn) ]]; then
-        sudo kill -9 $(pgrep openvpn)
-        sleep 1
-        if [[ -n $(command -v ip) ]]; then
-            local default="$(ip r | awk '/default/ {print $3}')"
-            while read -r route; do
-                sudo ip r d $route
-            done < <(ip r | tail -n +2 | \grep "via $default")
-            unset route
-        fi
+    info "Killing process and cleaning up..."
+    [[ -z $(pgrep openvpn) ]] || sudo kill -9 $(pgrep openvpn)
+    sleep 1
+    if [[ -n $(command -v ip) ]]; then
+        local default="$(ip r | awk '/default/ {print $3}')"
+        while read -r route; do
+            sudo ip r d $route
+        done < <(ip r | tail -n +2 | grep "via $default"); unset route
     fi
     rm -f creds.txt
-    echo "done"
+    info "done"
 }
 
 usage() {
@@ -141,23 +143,27 @@ usage() {
     echo
     echo "Options:"
     echo "    -h, --help      Display this help message"
+    echo "    --nocolor       Disable colorized output"
     echo "    -r, --random    Use random VPN gateway"
     echo
     exit $1
 }
 
+warn() { echo -e "${color:+\e[33m}[-] $@\e[0m"; }
+
 declare -a args
-unset conf gateway_selection
+unset conf gateway_selection help
+color="true"
 confdir="$HOME/.config/vpn"
-if [[ -f $confdir/vpn.conf ]]; then
-    conf="$HOME/.config/vpn/vpn.conf"
-fi
+[[ ! -f $confdir/vpn.conf ]] || conf="$confdir/vpn.conf"
 vpn="$(basename $(pwd))"
 
+# Parse command line options
 while [[ $# -gt 0 ]]; do
     case "$1" in
         "--") shift && args+=("$@") && break ;;
-        "-h"|"--help") usage 0 ;;
+        "-h"|"--help") help="true" ;;
+        "--nocolor") unset color ;;
         "-r"|"--random") gateway_selection="random" ;;
         *) args+=("$1") ;;
     esac
@@ -165,17 +171,12 @@ while [[ $# -gt 0 ]]; do
 done
 [[ -z ${args[@]} ]] || set -- "${args[@]}"
 
+# Check for valid params and missing dependencies
+[[ -z $help ]] || usage 0
 [[ $# -eq 1 ]] || usage 1
-
-if [[ -z $(command -v openvpn) ]]; then
-    echo "openvpn is not installed"
-    exit 2
-fi
-
-if [[ -z $(command -v jq) ]]; then
-    echo "jq is not installed"
-    exit 3
-fi
+for dep in jq openvpn; do
+    [[ -n $(command -v $dep) ]] || errx 3 "$dep is not installed"
+done; unset dep
 
 trap stop_vpn SIGINT
